dbus: refactor interfaces to be separate files for each
