+++ src/comm/dbus/control_interface.rs
@@ -10,69 +10,16 @@
 //
 // You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
 
-use crate::config;
-use crate::error::FpgadError;
-use crate::platforms::platform::{Fpga, OverlayHandler, Platform, new_platform};
+use crate::platforms::platform::Fpga;
+use crate::platforms::platform::OverlayHandler;
+use crate::platforms::platform::Platform;
+use crate::platforms::platform::new_platform;
+use crate::system_io::validate_device_handle;
 use log::trace;
-use std::path::{Path, PathBuf};
-use zbus::fdo;
-use zbus::interface;
+use std::path::Path;
+use zbus::{fdo, interface};
 
-pub struct StatusInterface {}
 pub struct ControlInterface {}
-
-fn validate_device_handle(device_handle: &str) -> Result<(), FpgadError> {
-    if device_handle.is_empty() || !device_handle.is_ascii() {
-        return Err(FpgadError::Argument(format!(
-            "{} is invalid name for fpga device.\
-                fpga name must be compliant with sysfs rules.",
-            device_handle
-        )));
-    }
-    if !PathBuf::from(config::SYSFS_PREFIX)
-        .join(device_handle)
-        .exists()
-    {
-        return Err(FpgadError::Argument(format!(
-            "Device {} not found.",
-            device_handle
-        )));
-    };
-    Ok(())
-}
-
-#[interface(name = "com.canonical.fpgad.status")]
-impl StatusInterface {
-    async fn get_fpga_state(&self, device_handle: &str) -> Result<String, fdo::Error> {
-        trace!("get_fpga_state called with name: {}", device_handle);
-        validate_device_handle(device_handle)?;
-        Ok(new_platform(device_handle).fpga(device_handle)?.state()?)
-    }
-
-    async fn get_fpga_flags(&self, device_handle: &str) -> Result<String, fdo::Error> {
-        trace!("get_fpga_flags called with name: {}", device_handle);
-        validate_device_handle(device_handle)?;
-        Ok(new_platform(device_handle)
-            .fpga(device_handle)?
-            .flags()
-            .map(|flags| flags.to_string())?)
-    }
-
-    async fn get_overlay_status(
-        &self,
-        device_handle: &str,
-        overlay_handle: &str,
-    ) -> Result<String, fdo::Error> {
-        trace!(
-            "get_overlay_status called with device_handle: {device_handle} and overlay_handle:\
-             {overlay_handle}"
-        );
-        validate_device_handle(device_handle)?;
-        Ok(new_platform(device_handle)
-            .overlay_handler(overlay_handle)?
-            .status()?)
-    }
-}
 
 #[interface(name = "com.canonical.fpgad.control")]
 impl ControlInterface {
