+++ src/comm/dbus/mod.rs
@@ -10,4 +10,5 @@
 //
 // You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
 
-pub mod interfaces;
+pub mod control_interface;
+pub mod status_interface;
